Implementación de las clases de composite.expression

+++ src/main/java/pd/composite/expression/Expresion.java
@@ -2,7 +2,7 @@ package pd.composite.expression;
 
 public abstract class Expresion {
 
-	public abstract double operar();
+	public abstract int operar();
 	
 	public abstract String toString();
 }
